requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-pinecone>=0.0.3
openai>=1.12.0
pinecone-client>=3.0.0
tiktoken>=0.5.2

# API and Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Caching and performance
redis>=5.0.0
hiredis>=2.3.0
slowapi>=0.1.9

# Text processing
beautifulsoup4>=4.12.0
nltk>=3.8.0
sentence-transformers>=2.3.0

# For hybrid search
collections-extended>=2.0.0
rank-bm25>=0.2.2

# Voice generation
elevenlabs>=2.0.0
tiktoken>=0.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
locust>=2.17.0

# Monitoring
python-json-logger>=2.0.0

# Knowledge graph dependencies
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
networkx>=3.0

# Optional: for development
ipython>=8.19.0
jupyter>=1.0.0
pdfplumber>=0.11.0
